% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertSparse}
\alias{convertSparse}
\title{S4 sparse matrix to arma::sp_mat}
\source{
\url{http://gallery.rcpp.org/articles/armadillo-sparse-matrix/}
}
\usage{
convertSparse(mat)
}
\arguments{
\item{mat}{S4 sparse matrix from the \code{\link[Matrix]{Matrix}}}
}
\value{
an arma::sp_mat object
}
\description{
Convert an S4 sparse matrix from the \code{\link[Matrix]{Matrix}} package
to an arma \code{sp_mat}.
}
