% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_graph.R
\name{.default_graph_param}
\alias{.default_graph_param}
\title{Generate data.frame with default vertex attributes}
\usage{
.default_graph_param()
}
\description{
Generate data.frame with default vertex attributes
}
