% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffuse.R
\name{diffuse}
\alias{diffuse}
\title{Diffuse scores on a network}
\usage{
diffuse(graph, scores, method, ...)
}
\arguments{
\item{graph}{\code{\link[igraph]{igraph}} object for the diffusion}

\item{scores}{list of score matrices. For a single input with a
single background, supply a list with a vector column}

\item{method}{character, one of \code{raw}, \code{gm},
\code{ml}, \code{z}, \code{mc}, \code{ber_s}, \code{ber_p}}

\item{...}{additional arguments for the diffusion method}
}
\value{
A list of scores, with the same length and
dimensions as \code{scores}
}
\description{
Function \code{diffuse} takes a network in
\code{\link[igraph]{igraph}} format and an initial state
to score all the nodes in the network.
}
