% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffuse_mc.R
\name{diffuse_mc}
\alias{diffuse_mc}
\title{Compute the heatrank using permutations}
\usage{
diffuse_mc(graph, scores, n.perm = 10000, diag.sum = 1,
  sample.prob = NULL, seed = 1, oneminusHeatRank = TRUE, K = NULL)
}
\arguments{
\item{graph}{igraph object}

\item{scores}{Recursive list}

\item{n.perm}{Numeric, number of permutations}

\item{diag.sum}{What to sum to the Laplacian diagonal?}

\item{sample.prob}{Numeric, probabilities (needn't be scaled) to permute the
input. This is passed to \code{\link[base]{sample}}'s \code{prob} argument.
If \code{NULL}, sampling is uniform.}

\item{seed}{Numeric, seed for random number generator}

\item{oneminusHeatRank}{Logical, should \code{1 - heatrank}
be returned instead of \code{heatrank}?}

\item{K}{Kernel matrix (if precomputed). If \code{K} is not supplied,
the regularised Laplacian will be computed on the fly and used.}
}
\value{
A list containing matrices of heatrank scores
}
\description{
Function \code{diffuse_mc} has an implemented parallelisation of the
Monte Carlo trials for diffusion in a (sparse) network.
}
