% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffuse_raw.R
\name{diffuse_raw}
\alias{diffuse_raw}
\title{Diffuse scores on a network}
\usage{
diffuse_raw(graph, scores, diag.sum = 1, z = FALSE, K = NULL)
}
\arguments{
\item{graph}{\code{\link[igraph]{igraph}} object for the diffusion}

\item{scores}{list of score matrices. For a single input with a
single background, supply a list with a vector column}

\item{diag.sum}{numeric constant or vector, amount
to add to regularise the Laplacian kernel if computed on the fly}

\item{z}{logical, should z-scores be computed instead of raw scores?}

\item{K}{optional matrix, precomputed diffusion kernel}
}
\value{
A list of scores, with the same length and
dimensions as \code{scores}
}
\description{
Function \code{diffuse} takes a network in
\code{\link[igraph]{igraph}} format and an initial state
to score all the nodes in the network.
}
