% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_graph.R
\name{generate_graph}
\alias{generate_graph}
\title{Generate a random graph}
\usage{
generate_graph(fun_gen, param_gen, class_prop,
  class_attr = .default_graph_param(),
  fun_curate = .connect_undirected_graph)
}
\arguments{
\item{fun_gen}{function to generate the graphs. Typically from
\code{\link[igraph]{igraph}}, like \code{\link[igraph]{barabasi.game}},
\code{\link[igraph]{watts.strogatz.game}},
\code{\link[igraph]{erdos.renyi.game}},
\code{\link[igraph]{make_lattice}}, etc.}

\item{param_gen}{list with parameters to pass to \code{fun_gen}}

\item{class_prop}{named character, proportions of each vertex class
(is internally normalised to sum up one). Default class names are
\code{c("source", "filler", "end")}.}

\item{class_attr}{data.frame with vertex classes as rownames and a column
for each vertex attribute. The name of the column will be used as the
attribute name.}

\item{fun_curate}{function to apply to the graph before returning it.
Can be set to \code{identity} or \code{NULL} to skip this step.
By default, the graph is connected: nodes not belonging to the
largest connected component are randomly wired to a node in it.}
}
\value{
An \code{\link[igraph]{igraph}} object
}
\description{
Function \code{generate_graph} generates a random network
using \code{\link[igraph]{igraph}} graph generators. Several models are
available, and
}
