% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_kernel}
\alias{is_kernel}
\title{Check if a matrix is a valid kernel}
\usage{
is_kernel(x, tol = 1e-08)
}
\arguments{
\item{x}{numeric, symmetric matrix to be checked}

\item{tol}{numeric, tolerance for zero eigenvalues}
}
\value{
scores in desired format
}
\description{
This function checks whether the eigenvalues are non-negative
}
\examples{
data(graph_toy)
K <- regularisedLaplacianKernel(graph_toy)
is_kernel(K)
is_kernel(K - 1)

}
