% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scores2shapes}
\alias{scores2shapes}
\title{Translate values into shapes}
\usage{
scores2shapes(x, shapes = c("circle", "square"))
}
\arguments{
\item{x}{numeric vector to generate shapes from}

\item{shapes}{character vector with two shapes, respectively
zeroes and ones}
}
\value{
Character vector with shapes
}
\description{
Translate 0/1 to shapes, by default \code{"circle"} and
\code{"square"}
}
\examples{
set.seed(1)
scores2shapes(rbinom(n = 20, size = 1, prob = .5))

}
