% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{serialHeatrank}
\alias{serialHeatrank}
\title{Compute heatrank for a single case}
\usage{
serialHeatrank(R, perm, G, ind)
}
\arguments{
\item{R}{dense matrix with the diffusion kernel}

\item{perm}{sparse matrix with the permutations}

\item{G}{sparse matrix with the heat sources}

\item{ind}{index of the G column for current source}
}
\value{
an arma::vec with node heatranks
}
\description{
The heatrank incorporates the correction \code{(r + 1)/(p + 1)}
instead of \code{r/p}
}
