% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sparsify2}
\alias{sparsify2}
\title{Sparsify arma::mat into arma::sp_mat}
\usage{
sparsify2(perm, nrow, header)
}
\arguments{
\item{perm}{dense matrix with the permutations}

\item{nrow}{number of rows for the sparse matrix}

\item{header}{Number of rows required from \code{perm} (will depend
on the size of the input list)}
}
\value{
an arma::sp_mat object
}
\description{
Return permutations as a 1-0 sparse matrix
}
