% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{to_list}
\alias{to_list}
\title{Convert input to list format}
\usage{
to_list(scores, dummy_column = "X1", dummy_list = "X1")
}
\arguments{
\item{scores}{object to reformat}

\item{dummy_column, dummy_list}{character, names for the dummy columns/items}
}
\value{
scores in list format
}
\description{
Convert any input to list format
}
\examples{
data(graph_toy)
x_v <- diffusion:::to_list(graph_toy$input_vec)
x_m <- diffusion:::to_list(graph_toy$input_mat)
}
