% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_graph.R
\name{.connect_undirected_graph}
\alias{.connect_undirected_graph}
\title{Function to connect a non connected graph}
\usage{
.connect_undirected_graph(g)
}
\arguments{
\item{g}{an igraph object}
}
\description{
Function to connect a non connected graph
}
