% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_graph.R
\docType{data}
\name{.default_graph_param}
\alias{.default_graph_param}
\alias{.default_prop}
\title{Generate data.frame with default vertex attributes}
\format{An object of class \code{numeric} of length 3.}
\usage{
.default_graph_param()

.default_prop
}
\value{
data.frame with default node class attributes

named numeric with default class proportions
}
\description{
Generate data.frame with default vertex attributes

Default proportions for randomly generated graphs
}
\keyword{datasets}
