% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffuse.R
\name{diffuse}
\alias{diffuse}
\title{Diffuse scores on a network}
\usage{
diffuse(graph, scores, method, ...)
}
\arguments{
\item{graph}{\code{\link[igraph]{igraph}} object for the diffusion}

\item{scores}{scores to be smoothed; either a named numeric vector,
a column-wise matrix whose rownames are nodes and colnames are
different scores, or a named list of such matrices.}

\item{method}{character, one of \code{raw}, \code{gm},
\code{ml}, \code{z}, \code{mc}, \code{ber_s}, \code{ber_p}}

\item{...}{additional arguments for the diffusion method}
}
\value{
A list of scores, with the same length and
dimensions as \code{scores}
}
\description{
Function \code{diffuse} takes a network in
\code{\link[igraph]{igraph}} format and an initial state
to score all the nodes in the network.
}
\details{
Input scores can be specified at three complexity levels.
A single set of scores to smooth can be represented as a named
numeric vector, whereas if several of these vectors that share
the node names need to be smoothed, they can be provided as
a column-wise matrix. However, if the number of unlabelled
entities varies from one case to another, a named list of such scores
matrices can be passed to this function. The input format will
be kept in the output.
A relevant note here is that \code{z} can give NaN for a particular node
when the observed nodes are disconnected from the node being scored.
This is because these nodes have
nor experimental neither network data.
}
\examples{
# Using a vector
data(graph_toy)
diff_scores <- diffuse(
    graph = graph_toy,
    scores = graph_toy$input_vec,
    method = "raw")

# Using a matrix
data(graph_toy)
diff_scores <- diffuse(
    graph = graph_toy,
    scores = graph_toy$input_mat,
    method = "raw")

# Using a list of matrices
data(graph_toy)
diff_scores <- diffuse(
    graph = graph_toy,
    scores = list(myScores1 = graph_toy$input_mat,
        myScores2 = graph_toy$input_mat),
    method = "raw")

}
