% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{package}
\name{diffuStats}
\alias{diffuStats}
\alias{diffuStats-package}
\title{diffuStats: an R package to compute and benchmark diffusion scores}
\description{
The \code{diffuStats} package consists of (i) functions to compute 
graph kernels, see \code{\link[diffuStats]{kernels}}, (ii)
the function \code{\link[diffuStats]{diffuse}} to compute the 
diffusion scores and (iii) the function 
\code{\link[diffuStats]{perf_eval}} and its wrapper 
\code{\link[diffuStats]{perf}} to compute performance measures. 
The user can find two vignettes in \code{browseVignettes("diffuStats")}: 
(1) a quick start with concise examples and (2) 
a detailed explanation of the 
implemented methods with a practical case study 
using a yeast protein dataset.
}
\author{
Sergio Picart-Armada <sergi.picart@upc.edu>, Alexandre Perera-Lluna
}
