% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{graph_toy}
\alias{graph_toy}
\title{Toy graph to play with diffusion}
\format{
An object of class \code{igraph} of length 10.
}
\usage{
graph_toy
}
\value{
An \pkg{igraph} object
}
\description{
Small graph that can easily be plotted and experimented with.
It has graphical parameters included, such as the vertex colour
and the layout. It also includes an example input. Has graph attributes
with example inputs and outputs, see \code{input_*} and \code{output_*}
from \code{list.graph.attributes(graph_toy)}
}
\keyword{datasets}
