% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{graph_toy}
\alias{graph_toy}
\title{Toy graph to play with diffusion}
\format{An object of class \code{igraph} of length 10.}
\usage{
graph_toy
}
\value{
An \code{\link[igraph]{igraph}} object
}
\description{
Small graph that can easily be plotted and experimented with.
It has graphical parameters included, such as the vertex colour
and the layout. It also includes an example input.
}
\keyword{datasets}
