% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{largest_cc}
\alias{largest_cc}
\title{Largest connected component}
\usage{
largest_cc(g)
}
\arguments{
\item{g}{\code{igraph} object}
}
\value{
A connected \code{igraph} object
}
\description{
Obtain the largest connected component of an \code{igraph} object
}
