% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sparsify2}
\alias{sparsify2}
\title{Sparsify arma::mat into arma::sp_mat}
\usage{
sparsify2(perm, nrow, G)
}
\arguments{
\item{perm}{dense matrix with the permutations}

\item{nrow}{number of rows for the sparse matrix}

\item{G}{sparse column matrix}
}
\value{
an arma::sp_mat object
}
\description{
Return permutations as a numeric sparse matrix
(can be binary or continuous)
}
