% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{to_x_from_list}
\alias{to_x_from_list}
\title{Convert list format to desired format}
\usage{
to_x_from_list(scores, x)
}
\arguments{
\item{scores}{list to reformat}

\item{x}{character, desired format}
}
\value{
scores in desired format
}
\description{
Convert any list format to the convenient one
}
\examples{
data(graph_toy)
x_v <- diffusion:::to_x_from_list(
    diffusion:::to_list(graph_toy$input_vec), "vector")
x_m <- diffusion:::to_x_from_list(
    diffusion:::to_list(graph_toy$input_vec), "matrix")
}
